backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pytest
pytest-asyncio
redis
openai